Added alerts for settings page, and user confirmation on delete

+++ src/app/components/settingspage/settingspage.component.ts
@@ -13,38 +13,55 @@ import {DatabaseService} from "../../services/database.service";
 })
 export class SettingspageComponent {
   database =inject(DatabaseService);
+  massButton = false;
 
   CreateDBFunction(){
     this.database.createDatabase().then(data => {
       console.log("Database Created Successfully")
+      alert("Created the DB!");
     }).catch(e => {
       console.error(e)
     });
   }
 
   CreateAllTablesDBFunction(){
+    this.massButton = true;
     this.CreateDroneTableDBFunction();
     this.CreateZoneTableDBFunction();
     this.CreateGroupTableDBFunction();
+    this.massButton = false;
+    alert("Created the DB Tables!");
   }
 
   DeleteDBFunction(){
-    this.database.deleteDatabase().then(data => {
-      console.log("Database deleted Successfully")
-    }).catch(e => {
-      console.error(e)
-    });
+    if(confirm("Are you sure to delete the Database?")){
+      this.database.deleteDatabase().then(data => {
+        console.log("Database deleted Successfully")
+        alert("The Database has been deleted!");
+      }).catch(e => {
+        console.error(e)
+      });
+    }
   }
 
   CreateDroneTableDBFunction(){
     this.database.CreateDroneTable();
+    if (!this.massButton){
+      alert("The Drone table has been created!");
+    }
   }
 
   CreateZoneTableDBFunction() {
     this.database.CreateZoneTable();
+    if (!this.massButton) {
+      alert("The Zone table has been created!");
+    }
   }
 
   CreateGroupTableDBFunction(){
     this.database.CreateGroupTable();
+    if (!this.massButton) {
+      alert("The Event table has been created!");
+    }
   }
 }
